requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
face_recognition
Pillow
python-jose
bcrypt
numpy
